feat(tokens): introduce DomModel for DOM interaction and selection management
- Added a new `DomModel` class that serves as a facade for resolving live DOM nodes, selections, and boundaries to model coordinates.
- Implemented methods for handling DOM node lookups, selection snapshots, and caret placement.
- Updated `index.ts` to export `SelectionSnapshot` from `DomModel`, consolidating selection-related types.
- Refactored `TokenModel` to utilize `DomModel` for DOM-related operations, enhancing separation of concerns.

This change enhances the architecture by encapsulating DOM interactions within the `DomModel`, promoting cleaner code and better maintainability.
