+++ packages/core/src/features/tokens/index.ts
@@ -4,6 +4,6 @@ export type {Token, TextToken, MarkToken, Markup} from './parser/types'
 export {annotate} from './parser/utils/annotate'
 export {denote} from './parser/utils/denote'
 export {TokenModel} from './model/TokenModel'
-export type {SelectionAnchor} from './model/TokenModel'
+export type {SelectionAnchor, SelectionSnapshot} from './DomModel'
 export {TokenHandle} from './model/TokenHandle'
 export {MarkController} from './MarkController'
